modify: draw studentTabs
